feat(dts-generator): introduce new directive for modules with named exports (#525)
The new .dtsgenrc option "modulesWithNamedExports" can be an array of
module names for which named exports are generated. For library modules,
this happens automatically, the y don't have to be listed.

The new option is useful for the webcomponent package files created by
ui5-tooling-modules. They should behave similar to library modules.